Disable getModulesPaths
